requirements.txt
httpx>=0.27.0
dnspython>=2.4.0
cryptography>=41.0.0
